Add files via upload
In this JavaScript Mini project, we will make a Digital clock that will display the current time in Hours, Minutes, and seconds. This JavaScript clock displays the time as per the Local time zone. We will use HTML and CSS to design this Clock, and for Updating the current time we will use JavaScript Date Object. You can use this Digital clock tutorial to make your Mini JavaScript Project for college or work portfolio.